refactor, add initial readme
