+++ index.ts
@@ -1,2 +1,1 @@
-export { RichTextBrowserParser } from "./src/parser/browser/RichTextBrowserParser"
-export { RichTextResolver } from "./src/resolver/RichTextResolver"
+export { RichTextBrowserParser, RichTextResolver } from "./src"
