+++ src/resolver/index.ts
@@ -1,0 +1,2 @@
+export * from './RichTextResolver';
+export * from './resolver-models';
