+++ src/resolver/resolver-models.ts
@@ -35,5 +35,9 @@ export interface IOutputResult<TOutput> {
 export interface IResolver<TInput, TOutput> {
     resolveAsync(input: TInput, resolvers: {
         resolveDomNode(domNode: IDomNode): Promise<TOutput>
-    }): Promise<IOutputResult<TOutput>>
+    }): Promise<IOutputResult<TOutput> | TOutput>
+}
+
+export interface ILinkFunction<TOutput> {
+    (node: IDomNode): TOutput;
 }
